medscript/requirements.txt
lark-parser==0.12.0
pytest==7.4.0 